tests/wpt/tests/webxr/layers/xrWebGLBinding_createCubeLayer.https.html

<!doctype html>
<title>XRWebGLBinding::createCubeLayer</title>
<link rel="help" href="https://immersive-web.github.io/layers/#dom-xrwebglbinding-createcubelayer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/webxr_util.js"></script>
<script src="../resources/webxr_test_constants.js"></script>
<script src="./xr_layer_promise_test.js"></script>

<canvas id="webgl-canvas"></canvas>

<script>

    function testCreateCubeLayer(xrSession, deviceController, t, { gl, xrBinding, xrSpace }) {
        return new Promise((resolve, reject) => {
            const valid_init = {
                space: xrSpace,
                viewPixelWidth: 1024,
                viewPixelHeight: 1024
            };

            t.step(() => {
                // Width and height must be equal.
                let invalid_size = Object.assign({}, valid_init, { viewPixelHeight: valid_init.viewPixelWidth + 1 });
                assert_throws_js(TypeError, () => xrBinding.createCubeLayer(invalid_size), "width and height are different");
            });

            // Test that a valid init works.
            t.step(() => {
                const layer = xrBinding.createCubeLayer(valid_init);
                assert_true(layer instanceof XRCubeLayer, "Valid init parameters must create an XRCubeLayer");
                resolve();
            });
        });
    }

    xr_layer_promise_test("Ensure XrWebGLBinding::createCubeLayer throws the appropriate errors",
        testCreateCubeLayer, TRACKED_IMMERSIVE_DEVICE, 'immersive-vr', { requiredFeatures: ['layers'] });
</script>
